requirements.txt
aspell-python-py3
num2words
pyicu
tqdm
